requirements.txt
flask
flask-smorest
python-dotenv
elasticsearch